requirements.txt
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
decorator==5.2.1
filelock==3.19.1
Flask==3.1.2
fsspec==2025.9.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.45.1
MarkupSafe==3.0.2
more-itertools==10.8.0
moviepy==2.2.1
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==2.3.3
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
pillow==11.3.0
proglog==0.1.12
pycparser==2.23
python-dotenv==1.1.1
pytorch-triton-rocm==3.5.0
regex==2025.9.18
requests==2.32.5
setuptools==70.2.0
sounddevice==0.5.2
soundfile==0.13.1
sympy==1.14.0
tiktoken==0.12.0
torch==2.9.0+rocm6.4
torchvision==0.24.0+rocm6.4
tqdm==4.67.1
triton==3.5.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
